fix(#1297): converge boolean check onto single bridge impl with null guards
- Add null guards to OCCTShapeBooleanCheckSingle and OCCTShapeBooleanCheckPair in OCCTBridge_Healing.mm
- Update isValidForBoolean and isValidForBoolean(with:) in Shape+Modeling.swift to forward to the new implementations
- Remove duplicate OCCTShapeBooleanCheck from OCCTBridge_Modeling.mm and its header
- Fix stale index entry in OCCTBridge.h umbrella header
- Merge BooleanCheckTests into BRepAlgoCheckTests (deleting BooleanCheckTests.swift)
- All gate scripts pass
- Tests pass
