+++ Sources/OCCTBridge/include/OCCTBridge_Modeling.h
@@ -1138,14 +1138,6 @@ OCCTShapeRef OCCTShapeEdgesToFaces(OCCTShapeRef compound, bool isOnlyPlane);
 /// @return Shape containing intersection edges/wires, or NULL on failure
 OCCTShapeRef OCCTShapeSection(OCCTShapeRef shape1, OCCTShapeRef shape2);
 
-// MARK: - Boolean Pre-Validation (v0.34.0)
-
-/// Check whether shapes are valid for boolean operations.
-/// @param shape1 First shape (argument)
-/// @param shape2 Second shape (tool), may be NULL for self-check
-/// @return true if shapes are valid for booleans
-bool OCCTShapeBooleanCheck(OCCTShapeRef shape1, OCCTShapeRef shape2);
-
 // MARK: - Split Shape by Wire (v0.34.0)
 
 /// Split faces of a shape by imprinting a wire onto a face.
